--- Uxcheckmate/BDD_Tests/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace BDD_Tests;
-
-public class Class1
-{
-
-}
