+++ Uxcheckmate/Moq_Tests/Controller_Tests/HomeController_Tests.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using Moq;
+using Moq.Protected;
+using NUnit.Framework;
+using Uxcheckmate_Main.Services;
+
+namespace Controller_Tests
+{
+    public class HomeController_Tests
+    {
+        //Set up
+        //Report
+        //index
+        //privacy
+        //ErrorPage
+        //Disclosure
+        //Guide
+    }
+}
