+++ Uxcheckmate/Moq_Tests/ErrorPage_Tests.cs
@@ -6,7 +6,7 @@ using Uxcheckmate_Main.Controllers;
 using Uxcheckmate_Main.Models;
 using Uxcheckmate_Main.Services; 
 
-namespace Moq_Tests
+namespace ErrorPage_Tests
 {
     public class ErrorPage_Tests
     {
