+++ Uxcheckmate/Moq_Tests/Service_Tests/ReportService_Tests.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using Moq;
+using Moq.Protected;
+using NUnit.Framework;
+using Uxcheckmate_Main.Services;
+
+namespace Service_Tests
+{
+    public class ReportService_Tests
+    {
+        
+    }
+}
