+++ Uxcheckmate/Uxcheckmate_Main/DAL/Abstract/IDesignIssueRepository.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
+
+using Uxcheckmate_Main.Models;
+
+namespace Uxcheckmate_Main.DAL.Abstract
+{
+    public interface IDesignIssueRepository : IRepository<DesignIssue>
+    {
+    }
+}
